完成home页面的基本框架

+++ src/main/java/com/example/entity/writing/Essay.java
@@ -18,7 +18,7 @@ import lombok.Data;
 public class Essay {
     int aid;
     String writer;
-    int Writingtime;
+    long Writingtime;
     String label;
     String photo;
     String background;
